Introduced a variable how_many to control
the number of packets to decode.
This value can be smaller or larger than
numSigs -- the number of distinct packets that
appeared in the input file.

Fewer (say 1) is often convenient for testing.

If you want to measure longer runs, specify
an arbitrary value for how_many -- the input
packets are simply recycled as input.
BUT, do NOT use this for BER (Bit Error Rate) evaluation.
For that you MUST have fresh input packets.