Device memory is allocated just once in an init phase.
To this properly, I should define a c++ class and
write an init and a cleanup function.
But, since I don't write c++, I made a hack to verify
the feasibility of this approach and placed what should all
be class local variables into globals.

This does work, and so before long I will convert this to a
class definition.