added file remapRowCols.cu
Swaps rows and cols in mapRows2Cols and mapCols2Rows
for better performance.

This is just a test -- which appears to be successful.
We get better performance, I think via more effective striding.

Once confirmed, I will simply supply the two maps with this ordering,
and drop the confusing remap.