Modifications to allow me to compare to the ocl code base
This is temporary and incomplete work.

The ocl code base uses the Min-Sum algorithm,
for which I have an implementation.
BUT, the subtraction (eta - lambda) is performed at different
(but equivalent) stages.
I modified this implementation to match the ocl version,
so that I could compare intermediate results.

WARNING.  so I modified cnpMinSum and bitEstimates accordingly.
I did NOT modify all the other cnp<name> implementations
to match with the modified bitEstimates here.

The ocl code base was restricted to regular codes --
those with the same number of checksPerBit for every check
(and bitsPerCheck for every bit).
But, no encoders were easily available for the codes used.
I used the "252" code  -  252 information bits rate 1/2
specifically for this comparison.
And, I did not bother to develop an encoder for this code.

I very simply read in a single encoded packet
(for which the information is alternativing 0 1, or even-odd)
and add random noise for each packet to decode.