Correction.
Original code had just thread 0 compute the product;
then each thread divided thread this product by the value
associated with the thread (removing itself from the product).

This should work -- but apparently did not perform exactly as I expected.

SINCE, all the threads are waiting on thread 0 to compute the product,
on a GPU it makes more sense for each thread to compute the product
not including itself.

Clearer AND more correct code.